% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cwt.R
\name{cwt}
\alias{cwt}
\title{Generate data from Continuous Wavelet Transforms}
\usage{
cwt(
  input,
  data = "Activity",
  out.bin = 15,
  t.cycle = 24,
  sleep.def = 5,
  low.per = 1,
  high.per = 35,
  rm.channels = c(),
  make.pval = FALSE,
  n.sim = 1000,
  method = "shuffle",
  boot = TRUE,
  boot.rep = 1000
)
}
\arguments{
\item{input}{Input data file. The input for this function must be the output of the function trimData(). See ??trimData().}

\item{data}{A choice between "Activity" and "Sleep" data to analyze using the wavelet transform.}

\item{out.bin}{Define the desired output bin size (in minutes). This defaults to 15.}

\item{t.cycle}{Define the period of the environmental cycle or a single day in hours. This defaults to 24.}

\item{sleep.def}{Definition of sleep. Traditionally, a single bout of sleep is defined as any duration of inactivity that is equal to or greater than 5-minutes. However, sometimes it may be of interest to examine longer bouts of sleep or specific bout durations; sleep.def allows users to change the definition of sleep. The default input is a single value vector of value 5. If users wish to analyse sleep only between 5 to 20 mins, the input must be c(5,20). Only relevant if data = "Sleep".}

\item{low.per}{Choose the lowest period (in hours) for analysis. This defaults to 1.}

\item{high.per}{Choose the highest period (in hours) for analysis. This defaults to 35.}

\item{rm.channels}{All the channels that users want to remove from their averaging. This must be a vector, i.e., channels must be separated by commas. For instance, if users choose to remove channels 1 to 5, 25 and 32, then the input should be either c(1,2,3,4,5,25,32) or c(1:5,25,32). This defaults to an empty vector, meaning no individuals are removed from analysis.}

\item{n.sim}{If TRUE, the number of simulations based on which p-values will be computed. Defaults to 1000.}

\item{method}{The method used for generating surrogate time-series against which p-values are generated. Also see help(analyze.wavelet). Available choices are "white.noise": white noise, "shuffle": shuffling the given time-series, "Fourier.rand": time-series with a similar spectrum, "AR": AR(p), "ARIMA": ARIMA(p,0,q). Defaults to "shuffle".}

\item{boot}{A logical (TRUE or FALSE). TRUE will bootstrap 95\% Confidence Intervals (CI) for the normalized amplitude for each period value. Defaults to TRUE.}

\item{boot.rep}{Number of replicates for bootstrapping 95\% CIs. Defaults to 1000. Only relevant if boot = TRUE.}

\item{n.days}{Number of days to analyse.}

\item{edge.rm}{A logical (TRUE or FALSE). TRUE will plot the scalogram with the edge effects removed. FALSE will plot the scalogram with the regions of the edge effect still retained in the plot. This defaults to FALSE.}

\item{make.p.val}{A logical (TRUE or FALSE). TRUE will generate p-values for the significance of detected periodicities. FALSE will not. Note that generating p-values significantly increases computation time. Defaults to FALSE.}
}
\value{
A \code{list} with two items:
\describe{
\item{CWT}{A \code{list} with all the details of the CWT computations for each fly, after removing dead or empty channels.}
\item{PeriodPower}{A \code{matrix} \code{array} with and 36 columns. The first columns stores period values. Columns 2 through 33 are for each fly. Column 34 is the averaged normalized amplitude (over flies) for each period value. Columns 35 and 36 store the lower and upper bounds of the 95\% CI.}
}
}
\description{
This function generates CWT results for each fly. Input for this function must be an output from the trimData() function. The output of this function is a large list. This function requires the package "WaveletComp".
}
\examples{
td <- trimData(data = df, start.date = "19 Dec 20", start.time = "21:00",
n.days = 1, bin = 1, t.cycle = 24)
cwt.dat <- cwt(data = td, boot = F)
}
