% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phaseCWTscalogram.R
\name{phaseCWTscalogram}
\alias{phaseCWTscalogram}
\title{Generate averaged scalograms from Continuous Wavelet Transforms}
\usage{
phaseCWTscalogram(
  input,
  data = "Activity",
  out.bin = 15,
  t.cycle = 24,
  n.days,
  sleep.def = 5,
  low.per = 1,
  high.per = 35,
  edge.rm = FALSE,
  make.pval = FALSE,
  n.sim = 1000,
  method = "shuffle",
  z.max = 0.5
)
}
\arguments{
\item{input}{Input data file. The input for this function must be the output of the function trimData(). See ??trimData().}

\item{data}{A choice between "Activity" and "Sleep" data to analyze using the wavelet transform.}

\item{out.bin}{Define the desired output bin size (in minutes). This defaults to 15.}

\item{t.cycle}{Define the period of the environmental cycle or a single day in hours. This defaults to 24.}

\item{n.days}{Number of days to analyse.}

\item{sleep.def}{Definition of sleep. Traditionally, a single bout of sleep is defined as any duration of inactivity that is equal to or greater than 5-minutes. However, sometimes it may be of interest to examine longer bouts of sleep or specific bout durations; sleep.def allows users to change the definition of sleep. The default input is a single value vector of value 5. If users wish to analyse sleep only between 5 to 20 mins, the input must be c(5,20). Only relevant if data = "Sleep".}

\item{low.per}{Choose the lowest period (in hours) for analysis. This defaults to 1.}

\item{high.per}{Choose the highest period (in hours) for analysis. This defaults to 35.}

\item{edge.rm}{A logical (TRUE or FALSE). TRUE will plot the scalogram with the edge effects removed. FALSE will plot the scalogram with the regions of the edge effect still retained in the plot. This defaults to FALSE.}

\item{make.pval}{A logical (TRUE or FALSE). TRUE will generate p-values for the significance of detected periodicities. FALSE will not. Note that generating p-values significantly increases computation time. Defaults to FALSE.}

\item{n.sim}{If TRUE, the number of simulations based on which p-values will be computed. Defaults to 1000.}

\item{method}{The method used for generating surrogate time-series against which p-values are generated. Also see help(analyze.wavelet). Available choices are "white.noise": white noise, "shuffle": shuffling the given time-series, "Fourier.rand": time-series with a similar spectrum, "AR": AR(p), "ARIMA": ARIMA(p,0,q). Defaults to "shuffle".}

\item{z.max}{The maximum color on the scalogram. Defaults to 0.5.}
}
\value{
A \code{plotly} \code{htmlwidget}.
}
\description{
This function generates a scalogram averaged over flies. The heatmap plots amplitude across the time-frequency domain. Input for this function must be an output from the trimData() function. The output of this function is a large plotly object. This function requires the packages "plotly" and "WaveletComp".
}
\examples{
td <- trimData(data = df, start.date = "19 Dec 20", start.time = "21:00",
n.days = 7, bin = 1, t.cycle = 24)
scalogram <- phaseCWTscalogram(input = td, n.days = 3, low.per = 1, high.per = 3)
}
